+++ vowpalwabbit/core/src/reductions/cb/cb_explore_adf_large_action_space.cc
@@ -43,7 +43,7 @@ public:
   {
     if (feature_value != 0.f)
     {
-      _triplets.emplace_back(Eigen::Triplet<float>(_row_index, index & _weights_mask, feature_value));
+      _triplets.emplace_back(Eigen::Triplet<float>(static_cast<int>(_row_index), static_cast<int>(index & _weights_mask), feature_value));
       if ((index & _weights_mask) > _max_col) { _max_col = (index & _weights_mask); }
     }
   }
@@ -295,7 +295,7 @@ std::shared_ptr<VW::LEARNER::learner> make_las_with_impl(VW::setup_base_i& stack
     bool apply_shrink_factor, size_t thread_pool_size, size_t block_size, size_t action_cache_slack,
     bool use_explicit_simd)
 {
-  uint64_t seed = (all.get_random_state()->get_random() + 1) * (uint64_t)10;
+  uint64_t seed = static_cast<uint64_t>((all.get_random_state()->get_random() + 1) * 10.f);
 
   auto data = VW::make_unique<cb_explore_adf_large_action_space<T, S>>(d, c, apply_shrink_factor, &all, seed,
       1 << all.initial_weights_config.num_bits, thread_pool_size, block_size, action_cache_slack, use_explicit_simd,
